--- src/main/java/org/scoula/domain/controller/ledger/entity/AccountCode.java
@@ -1,5 +1,0 @@
-package org.scoula.domain.controller.ledger.entity;
-
-public enum AccountCode {
-	WALLET_ASSET
-}
