--- src/main/java/org/scoula/domain/controller/ledger/entity/LedgerType.java
@@ -1,5 +1,0 @@
-package org.scoula.domain.controller.ledger.entity;
-
-public enum LedgerType {
-	DEBIT, CREDIT
-}
