--- src/main/java/org/scoula/domain/controller/ledger/entity/VoucherType.java
@@ -1,5 +1,0 @@
-package org.scoula.domain.controller.ledger.entity;
-
-public enum VoucherType {
-	CHARGE, TRANSFER
-}
