+++ src/main/java/org/scoula/domain/ledger/entity/AccountCode.java
@@ -1,0 +1,5 @@
+package org.scoula.domain.ledger.entity;
+
+public enum AccountCode {
+	WALLET_ASSET
+}
