+++ src/main/java/org/scoula/domain/ledger/entity/AccountType.java
@@ -1,4 +1,4 @@
-package org.scoula.domain.controller.ledger.entity;
+package org.scoula.domain.ledger.entity;
 
 public enum AccountType {
 	ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE
