+++ src/main/java/org/scoula/domain/ledger/entity/EntryType.java
@@ -1,0 +1,5 @@
+package org.scoula.domain.ledger.entity;
+
+public enum EntryType {
+	DEBIT, CREDIT
+}
