+++ src/main/java/org/scoula/domain/ledger/entity/LedgerType.java
@@ -1,0 +1,5 @@
+package org.scoula.domain.ledger.entity;
+
+public enum LedgerType {
+	DEBIT, CREDIT
+}
