+++ src/main/java/org/scoula/domain/ledger/entity/VoucherType.java
@@ -1,0 +1,5 @@
+package org.scoula.domain.ledger.entity;
+
+public enum VoucherType {
+	CHARGE, TRANSFER
+}
